Adding implementation of JSLint tool for Review Bot.
Parses all uploaded files to a review request and outputs descriptive
comments on the relevant .js source files on lines where style errors
are found. Options for this tool can be configured in the tool's admin
panel. This addition to ReviewBot requires that SpiderMonkey be installed
(other engines might also be suitable but have not been tested) to enable
the running of the jslint.js linting tool.